+++ src/ast/mod.rs
@@ -1,2 +1,3 @@
 pub mod expr;
 pub mod match_;
+pub mod decl;