+++ src/environment.rs
@@ -1,3 +1,4 @@
+use std::cell::RefCell;
 use std::fmt::Debug;
 use std::rc::Rc;
 use crate::ast::match_::Match;
@@ -9,38 +10,62 @@ where V: EnvVal {
     Entry {
         ident: String,
         value: V,
-        parent: Rc<Env<V>>
+        parent: Rc<RefCell<Env<V>>>
     }
 }
 
+pub type EnvWrapper<E> = Rc<RefCell<E>>;
+
 impl<V> Env<V> 
 where V: EnvVal {
-    pub fn empty() -> Rc<Self> {
-        Rc::new(Env::Empty)
+    pub fn empty() -> EnvWrapper<Self> {
+        Rc::new(RefCell::new(Env::Empty))
     }
     
-    pub fn associate(pattern: Match, value: V, parent: &Rc<Env<V>>) -> Result<Rc<Self>, String> {
+    pub fn associate(pattern: Match, value: V, parent: &EnvWrapper<Self>) -> Result<EnvWrapper<Self>, String> {
         let parent = Rc::clone(parent);
         value.unwrap_matches(&pattern).map(
             |matches|
             matches.into_iter().fold(
                 parent,
                 |parent, (ident, value)|
-                    Rc::new(Env::Entry { ident, value, parent }),
+                    Env::associate_ident(ident, value, parent),
             )
         )
     }
     
-    pub fn get(&self, ident: &str) -> Option<V> {
-        match self {
+    pub fn associate_ident(ident: String, value: V, parent: EnvWrapper<Self>) -> EnvWrapper<Self> {
+        Rc::new(RefCell::new(Env::Entry { ident, value, parent }))
+    }
+    
+    pub fn get(env: &EnvWrapper<Self>, ident: &str) -> Option<V> {
+        match *env.borrow() {
             Env::Empty => None,
-            Env::Entry { ident: id, value, parent } =>
+            Env::Entry { ident: ref id, ref value, ref parent } =>
                 if ident == id {
                     Some(value.clone())
                 } else {
-                    parent.get(ident)
+                    Env::get(parent, ident)
                 }
         }
+    }
+    
+    // Only should be used with declarations to solve recursive problem
+    // I strongly dislike the fact that I have to do this
+    // Also, write tests!!!
+    // This needs to work!!
+    pub fn set_value(env: &EnvWrapper<Self>, value: V) {
+        let mut inner_env = env.borrow_mut();
+        let new_inner_env = match *inner_env {
+            Env::Empty => Env::Empty,
+            Env::Entry {ref ident, ref parent, ..} =>
+                Env::Entry {
+                    ident: ident.to_string(),
+                    value,
+                    parent: Rc::clone(parent)
+                }
+        };
+        *inner_env = new_inner_env;
     }
 }
 
@@ -54,34 +79,34 @@ mod tests {
     
     basic_test! {
         no_parent_test
-        Env::associate(
+        Env::get(&Env::associate(
             Match::ident("a"),
             DummyValue::int("a", 1),
             &Env::empty()
-        ).unwrap().get("a") => Some(DummyValue::Int(1));
-        Env::associate(
+        ).unwrap(), "a") => Some(DummyValue::Int(1));
+        Env::get(&Env::associate(
             Match::ident("a"),
             DummyValue::int("a", 1),
             &Env::empty()
-        ).unwrap().get("b") => None
+        ).unwrap(), "b") => None
     }
     
     basic_test! {
         with_parent_test
-        Env::associate(Match::ident("a"), DummyValue::int("a", 1),
+        Env::get(&Env::associate(Match::ident("a"), DummyValue::int("a", 1),
             &Env::associate(Match::ident("b"), DummyValue::int("b", 2), &Env::empty()).unwrap())
-            .unwrap().get("a") => Some(DummyValue::Int(1));
-        Env::associate(Match::ident("a"), DummyValue::int("a", 1),
+            .unwrap(), "a") => Some(DummyValue::Int(1));
+        Env::get(&Env::associate(Match::ident("a"), DummyValue::int("a", 1),
             &Env::associate(Match::ident("b"), DummyValue::int("b", 2), &Env::empty()).unwrap())
-            .unwrap().get("b") => Some(DummyValue::Int(2));
-        Env::associate(Match::ident("a"), DummyValue::int("a", 1),
+            .unwrap(), "b") => Some(DummyValue::Int(2));
+        Env::get(&Env::associate(Match::ident("a"), DummyValue::int("a", 1),
             &Env::associate(Match::ident("b"), DummyValue::int("b", 2), &Env::empty()).unwrap())
-            .unwrap().get("c") => None
+            .unwrap(), "c") => None
     }
     
     basic_test! {
         with_tuple_test
-        Env::associate(
+        Env::get(&Env::associate(
             Match::tuple(
                 Match::ident("a"),
                 Match::ident("b")
@@ -91,8 +116,8 @@ mod tests {
                 ("b", 2)
             ]),
             &Env::empty()
-        ).unwrap().get("a") => Some(DummyValue::Int(1));
-        Env::associate(
+        ).unwrap(), "a") => Some(DummyValue::Int(1));
+        Env::get(&Env::associate(
             Match::tuple(
                 Match::ident("a"),
                 Match::ident("b")
@@ -102,7 +127,21 @@ mod tests {
                 ("b", 2)
             ]),
             &Env::empty()
-        ).unwrap().get("b") => Some(DummyValue::Int(2))
+        ).unwrap(), "b") => Some(DummyValue::Int(2))
+    }
+    
+    #[test]
+    fn setter_test() {
+        let expected = Some(DummyValue::Int(2));
+        let env = Env::associate_ident(
+                "a".to_string(),
+                DummyValue::Int(1),
+                Env::empty()
+        );
+        Env::set_value(&env, DummyValue::Int(2));
+        let actual = Env::get(&env, "a");
+        
+        assert_eq!(expected, actual);
     }
     
     #[derive(Debug, Clone, PartialEq)]